app/requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2023.12.2
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
huggingface-hub==0.20.1
idna==3.6
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pydub==0.25.1
Pygments==2.17.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
rsa==4.9
sympy==1.12
torch==2.1.2
torchaudio==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
